Guide Needs serious cleanup @wip

+++ app/controllers/user_tags_controller.rb
@@ -9,10 +9,10 @@ class UserTagsController < ApplicationController
   end
 
   def create
-    if current_user != @user
-      tag_names = TagCleaner.clean(params[:tag_names])
-      current_user.add_tags(@user, tag_names)
-    end
+    #if current_user != @user
+    tag_names = TagCleaner.clean(params[:tag_names])
+    current_user.add_tags(@user, tag_names)
+    #end
 
     render json: tags_summary(@user, current_user)
   end
@@ -39,7 +39,7 @@ class UserTagsController < ApplicationController
 
   def destroy
     user_tag = current_user.user_tags.find_by_id(params[:id])
-    user_tag.destroy
+    user_tag.destroy if user_tag
     render json: tags_summary(@user, current_user)
   end
